# removing allocations from startwith