--- a/src/tools/Tools.h
+++ b/src/tools/Tools.h
@@ -100,7 +100,7 @@ public:
 /// output, and the text between them is considered as a single word. Only the
 /// outer parenthesis are processed, to allow nesting them.
 /// parlevel, if not NULL, is increased or decreased according to the number of opened/closed parenthesis
-  static std::vector<std::string> getWords(std::string_view line,const char* sep=NULL,int* parlevel=NULL,const char* parenthesis="{", const bool& delete_parenthesis=true);
+  static std::vector<std::string> getWords(std::string_view line,const char* sep=NULL,int* parlevel=NULL,const char* parenthesis="{", bool delete_parenthesis=true);
 /// Faster version
 /// This version does not parse parenthesis and operates on a preallocated small_vector of string_view's
   static void getWordsSimple(gch::small_vector<std::string_view> & words,std::string_view line);
@@ -111,7 +111,7 @@ public:
 /// resulting line into an array of words
   static bool getParsedLine(IFile&ifile,std::vector<std::string> & line, const bool trimcomments=true);
 /// compare two string in a case insensitive manner
-  static bool caseInSensStringCompare(const std::string & str1, const std::string &str2);
+  static bool caseInSensStringCompare(std::string_view str1, std::string_view str2);
 /// Convert a string to a double, reading it
   static bool convertNoexcept(const std::string & str,double & t);
 /// Convert a string to a long double, reading it