<!--
  Feel free to delete not relevant sections below
-->

##### Description

with these modifications I speed up the initialization, I am looking also at a solution that @gtribello mailed me, I will remove the draft after I tried to add that modification and verified that the tests passes

As you can see most of the time is passed in `CaseInsensitiveCompare` and in `startWith`:

before

![sc_TORSION_128000nat_500steps_6threads](https://github.com/user-attachments/assets/0e843474-d35c-45ba-b2c2-76f7b9419ee8)

after:

![sc_TORSION_128000nat_500steps_6threads_comp_stringview_findkw_only](https://github.com/user-attachments/assets/f9d7e6a9-8949-48d8-89f0-d5523e7cfd46)


The time of master are:
```
BENCH:  Kernel:      this
BENCH:  Input:       sc_cmTORSION_CPU_128000.dat
BENCH:                                                Cycles        Total      Average      Minimum      Maximum
BENCH:  A Initialization                                   1   571.519895   571.519895   571.519895   571.519895
BENCH:  B0 First step                                      1     0.024423     0.024423     0.024423     0.024423
BENCH:  B3 Calculation part 2                            499     8.278312     0.016590     0.016000     0.019506
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1   579.820944   579.820944   579.820944   579.820944
PLUMED: 1 Prepare dependencies                           500     0.001118     0.000002     0.000001     0.000008
PLUMED: 2 Sharing data                                   500     0.275068     0.000550     0.000287     0.000830
PLUMED: 3 Waiting for data                               500     0.001679     0.000003     0.000002     0.000008
PLUMED: 4 Calculating (forward loop)                     500     7.945487     0.015891     0.015347     0.020815
PLUMED: 5 Applying (backward loop)                       500     0.051883     0.000104     0.000092     0.002660
PLUMED: 6 Update                                         500     0.018839     0.000038     0.000030     0.000087
```

```
BENCH:  Kernel:      this
BENCH:  Input:       sc_cmTORSION_CPU_128000.dat
BENCH:                                                Cycles        Total      Average      Minimum      Maximum
BENCH:  A Initialization                                   1   335.808614   335.808614   335.808614   335.808614
BENCH:  B0 First step                                      1     0.040812     0.040812     0.040812     0.040812
BENCH:  B2 Calculation part 1                            299     5.884897     0.019682     0.018033     0.023406
BENCH:  B3 Calculation part 2                            200     3.919282     0.019596     0.018236     0.022326
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1   345.652010   345.652010   345.652010   345.652010
PLUMED: 1 Prepare dependencies                           500     0.001030     0.000002     0.000001     0.000011
PLUMED: 2 Sharing data                                   500     0.142328     0.000285     0.000213     0.002293
PLUMED: 3 Waiting for data                               500     0.001774     0.000004     0.000002     0.000023
PLUMED: 4 Calculating (forward loop)                     500     9.606632     0.019213     0.017621     0.031642
PLUMED: 5 Applying (backward loop)                       500     0.061383     0.000123     0.000090     0.006774
PLUMED: 6 Update                                         500     0.022337     0.000045     0.000032     0.000101

```


The input is this one:
```
c: TORSION ...
128000-3 quadruplets
...
s: SUM ARG=c PERIODIC=NO
PRINT ARG=s FILE=TORSIONsc_128000 FMT=%8.4f STRIDE=1

FLUSH STRIDE=1

```

##### Target release

<!-- please tell us where you would like your code to appear (e.g. v2.4): -->
I would like my code to appear in release __XXXXX__

##### Type of contribution

<!--
  Please select the type of your contribution among these:
  (Change [ ] to [X] to tick an option)
-->
- [x] changes to code or doc authored by PLUMED developers, or additions of code in the core or within the default modules
- [ ] changes to a module not authored by you
- [ ] new module contribution or edit of a module authored by you

##### Copyright

<!--
  In case you picked one of the first two choices
  MAKE SURE TO TICK ALSO THE FOLLOWING BOX
-->

- [ ] I agree to transfer the copyright of the code I have written to the PLUMED developers or to the author of the code I am modifying.

<!--
  In case you picked the third choice (new module authored by you)
  MAKE SURE TO TICK ALSO THE FOLLOWING BOX
-->

- [ ] the module I added or modified contains a `COPYRIGHT` file with the correct license information. Code should be released under an open source license. I also used the command `cd src && ./header.sh mymodulename` in order to make sure the headers of the module are correct. 

##### Tests

<!--
  Make sure these boxes are checked. For Travis-CI tests, you can wait for them
  to be completed monitoring this page after your pull request has been submitted:
  http://travis-ci.org/plumed/plumed2/pull_requests
-->

- [ ] I added a new regtest or modified an existing regtest to validate my changes.
- [ ] I verified that all regtests are passed successfully on [GitHub Actions](https://github.com/plumed/plumed2/actions).

<!--
  After your branch has been merged to the desired branch and then to plumed2/master, and after the
  plumed official manual has been updated, please check out the coverage scan at
  http://www.plumed.org/coverage-master
  In case your new features are not well covered, please try to add more complete regtests.
-->