@Iximiel 2025-07-23
the last commit trims a little less that 70s with this input:
```
BENCH:  
BENCH:  Kernel:      this
BENCH:  Input:       sc_cmTORSION_CPU_128000.dat
BENCH:                                                Cycles        Total      Average      Minimum      Maximum
BENCH:  A Initialization                                   1   268.943159   268.943159   268.943159   268.943159
BENCH:  B0 First step                                      1     0.036019     0.036019     0.036019     0.036019
BENCH:  B2 Calculation part 1                            299     5.102488     0.017065     0.016090     0.023391
BENCH:  B3 Calculation part 2                            200     3.344050     0.016720     0.016154     0.019569
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1   277.423862   277.423862   277.423862   277.423862
PLUMED: 1 Prepare dependencies                           500     0.001217     0.000002     0.000001     0.000013
PLUMED: 2 Sharing data                                   500     0.260151     0.000520     0.000299     0.000860
PLUMED: 3 Waiting for data                               500     0.002131     0.000004     0.000003     0.000021
PLUMED: 4 Calculating (forward loop)                     500     8.132158     0.016264     0.015434     0.028601
PLUMED: 5 Applying (backward loop)                       500     0.056777     0.000114     0.000091     0.006444
PLUMED: 6 Update                                         500     0.020654     0.000041     0.000028     0.000154

```
![sc_TORSION_128000nat_500steps_6threads_comp_stringview_newgetkey_sw_insensitive](https://github.com/user-attachments/assets/f1b2050f-bd42-4d25-be6a-ec28b9f78adf)