@Iximiel 2025-07-30
#1307 supersedes this PR, and there the initialization passes from 600s to between 0.7 and 1.1 :)